pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "context-aware-whisper"
version = "0.1.0"
description = "Cross-platform speech-to-text using hotkey triggers and Groq Whisper API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Sukhdeep Singh", email = "sukhdeep.bangar@outlook.com"}
]
keywords = ["speech-to-text", "whisper", "transcription", "voice-typing", "cross-platform"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    # Audio recording
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    # Transcription API
    "groq>=0.4.0",
    # Clipboard operations (cross-platform)
    "pyperclip>=1.8.2",
    # Hotkey detection and typing (Windows/Linux)
    "pynput>=1.7.6",
    # Environment management
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# macOS-specific dependencies for Fn key detection and AirPods mute detection
macos = [
    "pyobjc-core>=9.0",
    "pyobjc-framework-Cocoa>=9.0",
    "pyobjc-framework-AVFoundation>=9.0",
    "pyobjc-framework-Quartz>=9.0",
    "rumps>=0.4.0",
]
# Local transcription using whisper.cpp (offline, private)
local = [
    "pywhispercpp>=1.2.0",
]
# Local LLM for aggressive text cleanup (MLX on Apple Silicon)
local-llm = [
    "mlx>=0.5.0",
    "mlx-lm>=0.4.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "hypothesis>=6.92.0",
]
# All optional dependencies for development
all = [
    "context-aware-whisper[macos,local,local-llm,dev]",
]

[project.scripts]
context-aware-whisper = "main:main"
caw-models = "context_aware_whisper.model_manager:main"

[project.urls]
Homepage = "https://github.com/sukhdeepsingh/context-aware-whisper"
Repository = "https://github.com/sukhdeepsingh/context-aware-whisper"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (may require hardware)",
    "requires_microphone: marks tests requiring microphone access",
    "requires_whisper: marks tests requiring whisper.cpp model",
    "requires_macos: marks tests requiring macOS-specific features",
    "requires_accessibility: marks tests requiring accessibility permissions",
    "slow: marks tests as slow (> 5 seconds)",
]
addopts = "-v --tb=short"
